# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sample App API swagger documentation",
        "description": "This is a sample API Documentation",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "apiteam@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/portfolio/tile/content/content/create": {
            "post": {
                "tags": [
                    "content_option"
                ],
                "summary": "Creates content option",
                "operationId": "ContentControllerCreate",
                "parameters": [
                    {
                        "name": "value_pl",
                        "in": "query",
                        "description": "Translation in polish language",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "value_en",
                        "in": "query",
                        "description": "Translation in english language",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "value_ru",
                        "in": "query",
                        "description": "Translation in russian language",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "tile_content_id",
                        "in": "query",
                        "description": "Card content id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "tile_content_shared_id",
                        "in": "query",
                        "description": "Card content shared_id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "tile_shared_id",
                        "in": "query",
                        "description": "Card shared id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/portfolio/tile/content/content/update": {
            "post": {
                "tags": [
                    "content_option"
                ],
                "summary": "Updates content option",
                "operationId": "ContentControllerUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Content option id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "shared_id",
                        "in": "query",
                        "description": "Content option shared id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "value_pl",
                        "in": "query",
                        "description": "Translation in polish language",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "value_en",
                        "in": "query",
                        "description": "Translation in english language",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "value_ru",
                        "in": "query",
                        "description": "Translation in russian language",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "selected",
                        "in": "query",
                        "description": "Is option selected?",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "position",
                        "in": "query",
                        "description": "Content option position",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "admin_visibility",
                        "in": "query",
                        "description": "Admin visibility",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "user_visibility",
                        "in": "query",
                        "description": "User visibility",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "tile_content_id",
                        "in": "query",
                        "description": "Card content id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "tile_content_shared_id",
                        "in": "query",
                        "description": "Card content shared id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/portfolio/tile/content/content/destroy": {
            "post": {
                "tags": [
                    "content_option"
                ],
                "summary": "Deletes specific card content option",
                "operationId": "ContentConteollerDestroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Content option id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "shared_id",
                        "in": "query",
                        "description": "Content option shared id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "tile_content_id",
                        "in": "query",
                        "description": "Card content option id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "tile_content_shared_id",
                        "in": "query",
                        "description": "Card content option shared id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/portfolio/tile/content/content/visibility/update": {
            "post": {
                "tags": [
                    "content_option"
                ],
                "summary": "Changes card content option admin or user visibility",
                "operationId": "ContentControllerUpdateVisibility",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Card content option id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "shared_id",
                        "in": "query",
                        "description": "Card content option shared id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "field",
                        "in": "query",
                        "description": "Admin or user visibility in:admin,user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "visibility",
                        "in": "query",
                        "description": "Visibility in:true,false",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/portfolio/tile/content/content/selected/update": {
            "post": {
                "tags": [
                    "content_option"
                ],
                "summary": "Changes card content option admin or user visibility",
                "operationId": "ContentControllerUpdateSelected",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Card content option id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "selected",
                        "in": "query",
                        "description": "Is card content option selected?",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/portfolio/tile/content/content/value/update": {
            "post": {
                "tags": [
                    "content_option"
                ],
                "summary": "Changes card content option admin or user visibility",
                "operationId": "ContentControllerUpdateValue",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Card content option id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "value_pl",
                        "in": "query",
                        "description": "Translation in polish language",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "value_en",
                        "in": "query",
                        "description": "Translation in english language",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "value_ru",
                        "in": "query",
                        "description": "Translation in russian language",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}